Refactor cookie handling code to make upcoming changes easier
Extract cookie information into an enum that gets stored as a field on
the SessionMiddleware. This will let us have another variant for the
Authorization: Bearer header, and should also let us configure the
cookie name in a subsequent commit. For now, just set a default value to
make the SessionMiddleware work the same way as it has.