Add example code to show how to authorize multiple routes
By extracting the authorization checking logic into a middleware then
using that middleware with a set of routes, this will reduce
duplication.

The logic within the middleware will likely be custom to each
application, so this isn't a great candidate for inclusion into
nickel-jwt-session proper, but it's useful to demonstrate!